Updated openstack/openstack Project: openstack/keystone  dd3a99f73776ea0edb900bdb8f1e90a0fd1f3009
Make fernet work with oauth1 authentication

Previously, fernet didn't know how to handle oauth1 authentication
flows. This patch adds a new token version to the fernet provider and
allows it to issue oauth1 tokens.

This work is being done so that we can get fernet to be feature
equivalent with the uuid token provider. Then we will be slightly
closer to making fernet the default token provider in keystone.

Closes-Bug: 1534252

Change-Id: I638404952597bb23dff01f80efb728b653e5560c