Updated openstack/openstack Project: openstack/keystone  7a60566c7d6945880114c86c4389be241f800083
Moves policy setup into a fixture.

The original implemention worked only because of the ordering of the
tests. By the time the tests that needed policy.json to be loaded ran it
had already been properly loaded. When running certain tests in
isolation the policy is not propertly setup, leading to a test failure.

Closes-Bug: #1520383
Change-Id: Icd041eb4ed8ddd580f49b4709ca5f05ab7315292