Updated openstack/openstack Project: openstack/keystone  98046501e4dffb7b9bf10226e0dbf5242721a76e
Add backend support for deleting a projects list

Adds a resource backend call that allows deleting a given list of
projects at backend level in an atomic operation. If any provided
project_id does not exist, it is ignored by the backend.

Subsequent patches will build on this to execute cascade deletion of a
project's hierarchy.

Co-Authored-By: Paulo Ewerton <pauloewerton@lsd.ufcg.edu.br>

Change-Id: I04c95acce9d355e96901e0c9f6675e0e0b959afe
Partially-Implements: bp project-tree-deletion