Updated openstack/openstack Project: openstack-infra/project-config  446ef7f4b94756e8dc2eb8dddf5a7977ce232919
Force the gate-irc-meetings-tox-ical postbuildscript to always run

The ical tox env on the irc-meetings repo will generate an index.html.
If it does generate an index.html and then fail to complete the tox-ical
job, the postbuild script will not be run. If the postbuild script
doesn't run there will be an index.html in the output directory, which
will be used instead of the apache DirectoryIndex module.  This will
obsure console.html makeing debbuging of the failure harder than it
needs to be.

To fix that for the postbuildscript to always run, which will rename
index.html to site-index.html if it exists.

Change-Id: I7c240bf2f757f668751e76eea439e63398a3c904