Updated openstack/openstack Project: openstack/keystone  9419fca5205974e98a019a60127253d365de9931
Fix release note of removal of v2.0 trusts support

The release note says it isn't possible to *obtain* a trust scoped
token with v2.0 anymore. However it was never possible anyways.

It should say it isn't possible to *validate* a trust scoped token
against v2.0 anymore. This patch fixes this.

Change-Id: I71d1f60c3d54b18cdbeb3047bc860c7906a42d48