Updated openstack/openstack Project: openstack/neutron  6ba5c89d10cd061efe79897861a2a92ce142e23b
ML2: delete_port on deadlock during binding

The previous logic was only catching mechanism driver exceptions so
it would leave behind a partially built port if a deadlock was
encountered during port binding.

The bug this closes was caused by a DBDeadlock being encountered when
a lock was attempted on the port binding record
(get_locked_port_and_binding). This would not be caught so the API
would retry the whole operation with the original created port
left behind. This resulted in two ports assigned to the same instance.

Closes-Bug: #1543880
Change-Id: I694a9d58002d72636225f99a5fb2b1ccc1cfb6e5