Updated openstack/openstack Project: openstack-infra/shade  661927392c245607f35ac111e7c5060d41cb4484
Use release version of Ansible for testing

We were using the unreleased (from git) version of Ansible for our
Ansible test job because at the time, 2.0 had not been released. Now
that it is released, use pip to install it so that we are less likely
to get broken by development changes to Ansible.

Change-Id: Ie714f42d26569044b6a3ee761bd5fb7451833d96