Updated openstack/openstack Project: openstack/neutron  8885f01dade434aa9e66ec5ea12a35c804dc19b9
Make netns_cleanup to purge resources of selected agent only

A new option --agent-type is introduced into netns_cleanup. When it is set
the utility processes namespaces for selected agent only. Possible values
are 'dhcp', 'l3', 'lbaas'. When the option is not set the behaviour stays
unchanged - all namespaces are processed.

Change-Id: Idaaffce767654bf21701a5329742f10a241b2cbf