Updated openstack/openstack Project: openstack/nova  b7deeddfac9f87b1a32816f7f552b95fb3fb0bc4
resource_providers, allocations and inventories models

Add migration and models for resource_providers, allocations
and inventories tables which will track availability and
consumption of generic resources.

Note that this also adds the uuid column to the compute_nodes table,
which is required for the later modeling. It's squashed in here
because it's part of this line of work, required for later, and
because fewer migrations == bettersauce.

Co-Authored-By: Dan Smith <dansmith@redhat.com>
Partially-Implements: blueprint resource-providers
Change-Id: Id77ec737201881ecba5233439d2a7c173471fcfa