Updated openstack/openstack Project: openstack-infra/project-config  b4a933c18f4791d297a42a7d6f4fcff71b48b4a5
Create stackviz element and script in nodepool

In order to get Stackviz running against tempest gate run data, it
will need to be pulled in its pre-built form (i.e. with npm deps
installed) from the nodepool images. This change adds a couple steps
to build Stackviz so that it is in a ready state for retrieval and use
on the logs server.

A future devstack-gate patch will pull the stackviz directory from the
nodepool images onto the logs server. This nodepool patch is required
first in order to prepare stackviz for download. Installing the npm
modules takes a considerable amount of time (~5-10 mins), so it is
best that this work is done on the images before gate jobs are run.

Change-Id: I9bc50b8eba696264d2b97ed20a37f28f4768abb8