Updated openstack/openstack Project: openstack/python-cinderclient  3a170421286288d828682e6b92123c8061f9e9a3
Put py34 first in the env order of tox

To solve the problem of "db type could
not be determined" on py34 we have to run first the py34 env to, then, run
py27. This patch puts py34 first on the tox.ini list of envs to avoid this
problem to happen.

"Closes-Bug: #1533919"
Change-Id: Ib247b4000d9c1d906ef55fa2b5cc990cf1120850