Updated openstack/openstack Project: openstack/cinder  1a8fad17bf3a28f8bec5fb0d5238568e035737e0
Fix image volume creation error

When "image_upload_use_cinder_backend = True" is specified to
the volume backend and the Glance cinder store is enabled,
upload-to-image command fails with the following error:

  Create clone_image_volume: xxx for image xxx failed
  (Exception: An object of type VolumeAttachmentList is required
   in field volume_attachment, not a )

This is due to unnecessary attributes specified in the volume
params, such as 'volume_attachment' or 'consistencygroup'.
This patch removes the unnecessary attributes from the params.

Change-Id: Ia416cb6686312e1591de7fe40a7ce7d6aa2fce7d
Closes-Bug: #1529762
Related-Bug: #1528087