Updated openstack/openstack Project: openstack/oslo.messaging  eb4df4ecd061bb8f1c6c295583e7888ac304aed3
Remove server queue creating if target's server is empty

If target passed to MessageHandlerServer doesn't contain server,
It is not needed to create queue for server.
Also this patch removes mandatory flag for sending message during fanout
cast because is is possible to cast fanout messages when nobody listen
queue and this situation is ok.

Change-Id: I4f8930e7b939c62bc1e97b4acaba6884fe1b97b2