Updated openstack/openstack Project: openstack/keystonemiddleware  310a5b204e43174630c70b1e509458f0a4228dfc
Remove except Exception handler

Our except Exception handler catches any passing exception and turns it
into a webob 500 error. However the default behaviour of uncaught
exceptions is to be rendered as 500 errors with actual debug output so
all we are doing is masking useful failure stack traces.

Removed the only exception that triggered this case. For offline token
validation we really should know all the types of exception that can be
raised and handle them.

Change-Id: I89d84a60b76c7a7e08e3eb453b1880b064983fbf