Updated openstack/openstack Project: openstack/heat  14f3bb75a36d082795d295f4098392ed93720c97
Set DB max_overflow to match thread pool size

Set max_overflow config option of database to match the number of
workers that can be spawned simultaneously. The maximum number of RPC
workers that can be spawned is determined by executor_thread_pool_size
config option.

This ensures there won't be lack of DB sessions when many worker
threads/eventlets are spawned when the engine is under load.

The DB connections that are part of overflow bucket are closed
automatically when they are not in use, so heat doesn't run out of
'lsof' limit.

Closes-Bug: #1491185
Change-Id: I4fcfe50b53b4d476f7483bb860667c42d0c54a19