Updated openstack/openstack Project: openstack-infra/project-config  584164cad9664566fe8a98b51d599f8f6db722d1
Add job gate-swift-tox-func-in-process-fast-post

Adds non-voting gate-swift-tox-func-in-process-fast-post.

The new 'func-in-process-fast-post' tox env tests
swift using an alternative configuration to increase
test coverage. A new env is used rather than changing
the definition of the current 'func' env because some
developers use the 'func' env for local testing.

The new job is non-voting until it is proven stable.

Depends-On: Ia37f6df1dc38e44ef7404fbf0a52f6fc22fae0c2
Change-Id: Ibdd5f8a8201ce377e73f88892d7a42c81d1f5d8d