Updated openstack/openstack Project: openstack-infra/project-config  252d4e3e1911a97fa368e24672dc3a0d6c5b5fb1
import release-test repository

The release team wants to have a repository to test changes to the
release tools against, instead of using real releases from real
projects. This repo will serve as a dummy project that we can use for
testing releases, announcements, etc. without risking creating a bad
release of something real.

The permissions are set to let anyone with access to the release-tools
repository to modify this new repository. We don't need a new gerrit ACL
group for this repo.

Change-Id: I5c490dbd660bcea29b2b655d9f2e8131c346ecdb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>