Updated openstack/openstack Project: openstack/sahara  076940aa43daf2be700c99902c1a98e9b1f944dc
Remove overlap of 'is_default' and 'is_protected' for templates

The 'is_default' field remains as a marker for default templates
so that they can be managed from the sahara-templates tool. However,
the 'is_default' field will no longer protect a default template
from deletion or update. The 'is_protected' tool should be used
for that. To ensure that existing default templates have the
same protection, a migration script is included that sets
'is_protected' on any curret default template.

Change-Id: I1d77eb7fe104821ad44ddb972d0bec383ee5b8de
Partial-implements: blueprint replace-is-default