Updated openstack/openstack Project: openstack/horizon  6872355d178cebbdafe5fc433d45ff9bfb9c94ed
Remove current consumer in QoS spec edit

In QoS spec edit consumer, the choices including the current choices.
If we select the current choice again, we are getting form error.
Instead of this, this patch removing the current choice from the list.
Displays the current consumer in read only field.

Change-Id: Ie30517a34146f9bf1e8859088cb030f4032e6649
Closes-Bug: #1466773