Updated openstack/openstack Project: openstack-infra/puppet-elasticsearch  a3745ab6ee5dbd03a98f123c29fdd6079ba26c7b
Make network.publish_host configurable

The network publish host value defaults to the ip address of the
first non-loopback interface, but this is not always an appropriate
value.

In some cases this value may need to be set to something other than
the ip address of the local interface, such as when connections should
be made to a floating ip address.

Change-Id: I9ace9bc80607e8d60478e869ef13feea8c49e10a