Updated openstack/openstack Project: openstack/neutron-vpnaas  c547e2abc6d0311bae51fc8eff330dcb22be561a
Add VPNaaS API tests in neutron-vpnaas tree

Change [1] implemented that VPN tests are removed from
neutron tree. This change adds the same test with their
resources to run them in neutron-vpnaas tree.

In the other patch, a new CI job can be pushed in project-config
so that this api test can be tested.

[1] https://review.openstack.org/278606

Change-Id: Ia7d7b53aaeda8c328dcb7694f1a9f974e4f2ca31
Closes-Bug: #1483417