Updated openstack/openstack Project: openstack/designate  2297a97c15dfa50d082d35a8a6ce53557bfa81c0
Update numbering of dashboard enable files

Recently, the Horizon Dashboard project launched a new approach
for loading dashboard panels at startup, and along with that, a
new 4 digit numbering scheme for the files that enable that to
happen.

The change is happening in the designate-dashboard project. This
patch updates the file names that get sym-linked into Horizon.

Change-Id: Ifc9ee0a2afbd6b826993ff03c1d23aacabf3144e