Updated openstack/openstack Project: openstack/oslo.log  6e7e77bf389637f996ba0e6d099d11983b02c5ad
stop making a copy of options discovered by config generator

In order for the config generator's new hook for updating default values
to work, we have to return the same Opt objects from list_opts() that
our set_defaults() function is going to modify. There's no real need to
be making copies of the objects anyway, so just stop doing it.

Change-Id: I9373ce443c8199b6ccc43dc8ca2d51c7d2f0a084
Signed-off-by: Doug Hellmann <doug@doughellmann.com>