Updated openstack/openstack Project: openstack/oslo.db  1e9bcbf17d23cc602f0e3c4d31a1f7407907f0c6
stop making a copy of options discovered by config generator

In order for the config generator's new hook for updating default values
to work, we have to return the same Opt objects from list_opts() that
our set_defaults() function is going to modify. There's no real need to
be making copies of the objects anyway, so just stop doing it.

Change-Id: I62c2f733f1e7943449aa744ed15dac4addb3e1f7
Signed-off-by: Doug Hellmann <doug@doughellmann.com>