Updated openstack/openstack Project: openstack/diskimage-builder  6f6a096d94f5f9d46bc05572bbe31c5af63c6d7d
Handle install with pip -e

As described in the comments, inspect the installation to see if we
have been installed with "pip -e" and, if so, make sure we reference
the scripts from the source location rather than the
system-installations.

Update the documentation with a terse but helpful quick-start to show
an easy way to start developing a change using this.

Closes-Bug: #1491035
Change-Id: I0460061b834a2b854175f8c9be2be8d38c540c9d