Updated openstack/openstack Project: openstack/nova  7616c88ad3a2769e9c9ee8a51ac55ddeed0bfd84
Change populate_security_groups to return a SecurityGroupList

populate_security_groups expects an instance object to be passed in and
just sets security_groups on it.  An upcoming change needs this method
to return the SecurityGroupList object instead because it will not be
getting set on the instance at this point.  This simply changes the
method to return the object rather than setting it on the instance.

Change-Id: Ifa40fc061ea041cfcc229ee7255b3c084f3bba6f