Updated openstack/openstack Project: openstack/python-openstackclient  088cb82631fc0b7918e2dfcde96207efc819cad5
Allow custom log levels for other loggers

It would be convenient to be able to enable special logging for
various components that openstack uses.  The biggest thing is
the --debug prints a lot of information when often all I want to
see is the outgoing requests/responses.  To get just that logged
you would put this in your clouds.yaml:

    logging:
        keystoneclient.session: debug

Closes-Bug: #1484660
Change-Id: I15c2607e8262f10903dd831ee8622fb5d6315310