Updated openstack/openstack Project: openstack-infra/project-config  ce436c6988662ea7625adce45ee546ac191b12d8
Make propose_translation_update more robust

If there're no translations at all, the locale directory is empty
but we try to add content and that fails like:
 + git add 'ironic_ui/locale/*'
 | fatal: pathspec 'ironic_ui/locale/*' did not match any files

Do not use a wildcard, just add the directory. That way the git command
and will not fail - and the rest of the script works fine.

Change-Id: I2a67b692f437a1bced9507e66cce217765d5d913