Updated openstack/openstack Project: openstack/manila  8bf1ec14c711383c76ff677b182be7becb153da2
Avoid KeyError on instance_id in ensure_service_instance

There are cases where the server details don't have an instance_id
set but we need to check for that so we can avoid a KeyError. We
also log this case as a warning like the InstanceNotFound in the same
method. Ideally we'd log the share server ID for context but we don't
have that available in this method which is given ShareServerBackendDetails
key/values.

Change-Id: I8cabf7f3a5a3480a1f55a19fbdcebfc272906332
Closes-Bug: #1542365