Updated openstack/openstack Project: openstack/python-manilaclient  bb53633fd9b19d5323b5512209a4566a15a8a78f
Fix Mutable default argument

Python’s default arguments are evaluated once when
the function is defined, not each time the function is called.

This means that if you use a mutable default argument and
mutate it, you will and have mutated that object for all
future calls to the function as well.

more details about this wrong usage here:
http://docs.python-guide.org/en/latest/writing/gotchas/#mutable-default-arguments
Closes-Bug: #1534974

Change-Id: Ia87abe0d020d7840dbf5f8a3063f057bab92c2ba