Updated openstack/openstack Project: openstack/oslo.config  3195602efde47b5600c0f624bf29ff80e0148475
refactor generator._list_opts for further enhancement

The _list_opts() function combines the behavior for loading the options
and the behavior for cleaning and filtering them. This patch breaks it
up a bit and updates the relevant tests, to prepare for the next patch
which will add some new behavior to _list_opts().

Change-Id: Ie6616e76bf0c23a6303cdb1e441754ccba37a9c3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>