Updated openstack/openstack Project: openstack/oslo.config  1482e5c601d0f75a13972cc808c5fc5a98513624
add generator hook for apps to update option defaults

Provide a way for applications to hook into the config sample generator
to change the defaults for options before the sample output is
rendered.

We have to use the same public functions exposed by libraries
for this purpose, but we can't guarantee the import order of the option
discovery functions, so this patch adds an explicit hook to let the
application invoke set_defaults() before we ask the functions for their
options.

The results depend on the libraries returning the original Opt objects
rather than copies, so we may have to make changes to some list_opts()
implementations in other libraries.

Change-Id: Ieedf522eda1f57d6e4ea1b30a01df656019ac0c5
Signed-off-by: Doug Hellmann <doug@doughellmann.com>