Updated openstack/openstack Project: openstack/taskflow  6241a5c5920e5997fe1cc392413cc7ce4536c33c
Remove need for separate notify thread

Instead of having a periodic notification thread that
will drop messages to try to find workers we can just
have this same work be done in the periodically called
on_wait callback that is already used for expiring and
matching workers to new/updated workers.

This avoids having one more thread that doesn't do all
that much (and activating it during waiting calls will
be often enough to achieve its goal in life).

Change-Id: If80233d13d914f2ed3665001a27627b78e6ee780