Updated openstack/openstack Project: openstack-infra/project-config  9492e19920772f0e1f515b50a7c1a61d649e2543
Remove {name} jobs from devstack-jobs

The job group devstack-jobs is only used with the name devstack-jobs in
projects.yaml - but it should be with repository names. So, all the
jobs here are also setup separately.

So, remove these jobs from devstack-jobs, they are totally wrong here
and will not work as expected.

Change-Id: I55a29a280a66ffe4eb187249f58f24f3a722f294