Updated openstack/openstack Project: openstack/neutron  d51a18c4d09b541443f8ed3d504c586a685d3d66
Pecan: Fix association of plugins with resources

The process was leveraing the supported_extension_aliases
attribute, which might not provide a complete list of extension
aliases.

With this patch now Pecan leverages helper methods available
in the extension manager, which have been slightly refactored in
order to allow for retrieving aliases for a single plugin.

Closes-Bug: #1542277

Change-Id: Id944b27dcd84a92bf15242f5797408bf83b0d317