Updated openstack/openstack Project: openstack/ironic  6d8768e87c971f36a7abba32c2c57aba4de59489
Augmenting the hashing strategy

It was only 'sha1' algorithm being used till now in ``hash_file``
method to create hash of the file contents. Need to add 'md5'
to keep the consistency with that of IPA code base. Also, this
method now supports all the algorithms hashlib does. As this
method wasn't being used at all, therefore making the default
algorithm to 'md5'. Will be used in oob firmware update (manual
clean step) to verify the checksum of the images.

Change-Id: I8159fd64d1c074bd539817a40900e34f8fae30d4