Updated openstack/openstack Project: openstack/python-openstackclient  67bec569f866f8ffe2eac4d268fab6a610608e90
Rename parameter "identifier" to "network" in network commands

In other commands, the name or ID of an object is just the name
of the object. For example, name or ID of a server is "server",
router is "router". So, do not use "identifier" in network commands.

Also, the parameter in doc file network.rst is not "identifier",
but "network".

Change-Id: I1ec3beefbb878a207bca280b994ca176ef04ee2d