Updated openstack/openstack Project: openstack/neutron  2fbfd7dadc4c3fb4d01b299d36071143c9d40879
Get rid of UnionModel for RBAC

The union model approach was completely broken because it
didn't keep track of which model each result actually was.

This patch just strips it out and replaces get_rbac_policies
with queries to each model. This will mean pagination is broken
once multiple rbac types are in place, but everything else should
work fine.

Co-Authored-By: Haim Daniel <hdaniel@redhat.com>

Closes-Bug: #1542815
Change-Id: I1e91aa22d093d50e5a9d318f24d09bb65e072246