Updated openstack/openstack Project: openstack-infra/project-config  ec4ecf3e3f2ceaf6a0b1d86e8beef98ad4a9b710
Fix constraints for requirements repo

Fix zuul-git-prep-upper-constraints to not copy a non-existing file. If
requirements repo is to be checked out REQS_DIR is an empty directory,
and thus we need to special case it.

Note that if we check out requirements repo, we can use the
upper-constraints file from that checkout, so no need to copy anything
around.

Change-Id: I863de2c3d5847e9c1943bb25e28b86a68871693f