Updated openstack/openstack Project: openstack/nova  2e28de7c5b7b108d4c37a948302696b69ab9d542
Allocate UUID for compute node

This patch causes us to allocate a UUID for ComputeNode on create or
when loaded from the database. Since we don't currently have them,
query needs to handle this case to migrate us to the point where we
can add a uniqueness and non-null-ness constraint on that column in
the database. We take a small hit every time we query a compute node
that doesn't have a uuid, by generating it and saving it back to the
database immediately, but this only happens once.

Related to blueprint compute-node-inventory

Change-Id: I9e8bd5f2ed064923ed04feb864b42e4b2948812d