Updated openstack/openstack Project: openstack/nova  5aee67a80a30725a7d2b95533baf8bfb73476ef1
Move Disk allocation ratio to ResourceTracker

We already moved the CPU and RAM ratios to the compute nodes in Liberty
but we forgot to add the disk one. Adding it now.

That change specifically implies that nova.conf is consistent across the whole
cloud since what was formerly defined on the scheduler side will have its
definition by the compute nodes.

Co-Authored-By: Sylvain Bauza <sbauza@redhat.com>
UpgradeImpact

Change-Id: Id4b44dc4c3391563154a30406a5f1e7a2dc015b3
Partially-Implements: blueprint disk-allocation-ratio-to-rt