Updated openstack/openstack Project: openstack/heat  0746e81102e3412c57aef5976a4e9ed1e1a49bb8
Remove db refresh in Stack object get

When refreshing a Stack object, we first retrive the object from the
database (which handle NotFound errors) and then call refresh in a
different transaction. By that time, the db object can be gone and
refresh would fail, and it also doesn't bring any benefit, so let's
remove that call.

Change-Id: I63af65d2a409ddf4179672ebb054852a8732a1fe
Closes-Bug: #1543922