Updated openstack/openstack Project: openstack-dev/bashate  6d166755b3eab3e0f650657c74d35dd87908c4b8
Fix final newline testing

Since we don't use the fileinput loop since
I781c1b642eb15fc7e2914791d5b77e8a4752db79, we can just check the last
line of the file before we start reading the next file (using the
bulitin fileinput loop tells you when a new file starts, but doesn't
tell you went a file ends, hence this work-around).

Which is actually the correct thing to do, because as it is, you can't
detect the lack-of-newline unless you're checking two files.  This
removes a bunch of prev_* variables that aren't needed.  This is also
I think why we never had a test-case.  So fix that too.

Change-Id: I54963cd25efc9f9603fbcc60b8e25bb11ea2d7db