Updated openstack/openstack Project: openstack/zaqar  3746b4536246394d19e591e0bf1254fff03e0767
Ensure hmget return right data format

The Jenkins is done due to the funtional test[1] in redis backend.

The reason is that we use hmget to get the information but hmget
can't distinguish integer. it returns all value in string format.

This patch did a quick fix to let jenkins work.

[1]: https://github.com/openstack/python-zaqarclient/blob/master/zaqarclient/tests/queues/subscriptions.py#L212

Change-Id: I0dcb50535913457ae1ad705939dcff555c6e1ae8