Updated openstack/openstack Project: openstack/ceilometer  2bd126bb865427dce479892a3e8cbdde3ad248c4
Set None explicitly to filter options.

When user is set to None in query, string value 'None'
was set instead of None.

Note: This fix is for mysql.
Fix for mongo is provided in:
https://review.openstack.org/#/c/260331/

Note2: Here, test for resource_id is removed as,
in sqlalchemy implementation, resource_id cannot be None.

Change-Id: I4bcf9283912e857a9902c62df25d0d5090670b10
Partial-Bug: #1388680