Updated openstack/openstack Project: openstack/oslo.config  3dfd4a4e657426534b132cbe0c2dbed554d060f2
Fix wrong check with non-None value when format group

This patch is same to the bug-fix 'Fix can't generate config sample
with non-None value'[1].

There is pitfall in statement like:
  if getattr(opt.type, 'min', None)
The check returns False When opt.type.min equals 0,
We should check this with None.

This patch will fix it and add unit tests for the change.

[1]https://review.openstack.org/#/c/267566/2

Change-Id: Iec6d8d70c67f641ae53976bde13697ce0f4982c1