Updated openstack/openstack Project: openstack-infra/project-config  3d4bca2dc53eb57d03c8a14ea7bd919ea0188502
Exlude gate-tempest-dsvm-neutron-src-oslo.service from stable/liberty

This job is broken on stable/liberty because when installing oslo.service
from source on stable/liberty, the version is 0.9.1.dev6. upper-constraints
in stable/liberty has oslo.messaging===3.0.0 which requires
oslo.service>=0.12.0, but since we're just trying to test the oslo.service
change from source (0.9.1.dev6) we fail with a ContextualVersionConflict.

So for now just exclude the job from running on oslo.service stable/liberty
changes since it's fundamentally broken.

Alternatives to this are:

1. Lower the oslo.messaging upper-constraint bound in stable/liberty to
   2.5.0. That is the version that matches stable/liberty g-r for
   oslo.service>=0.7.0. However, upper-constraints is really saying that
   oslo.messaging 3.0.0 should work with the version of oslo.service that
   is in upper-constraints, which is oslo.service===1.1.0, and that would
   work, so lowering oslo.messaging in u-c for this case is hacking around
   that system.

2. We could release a 0.12.1 version of oslo.service from stable/liberty but
   that is also masking the issue and doesn't resolve the fact that you
   can't install oslo.service 0.9.0 and oslo.messaging 3.0.0 together in
   stable/liberty, even though that's what the min bound in g-r for
   oslo.service and the upper-bound in u-c for oslo.messaging says you can
   at least attempt to do.

Ideally, if it were up to me, I'd just cap oslo.messaging<2.6.0 in g-r on
stable/liberty and be done with it, but that's a bigger discussion to sort
out.

Change-Id: Ic20f69c76d153d9d4a4f4c229ac3aed21b6db8f5
Related-Bug: #1529594