Updated openstack/openstack Project: openstack/nova  dbdf27021cc23bf8cec8d3f51fa03c6dac5da53a
objects: Rename PciDevice _migrate_parent_addr  method

We want to make this part of the API the framework for building
nova-manage scripts hooks into, so we want to make sure the name
reflects this. We also add a no-op method to the NovaObjects base class
to emphasize this.

In the future we may want to add more knobs to this interface but that
should be easily doable since that method is not remotable.

Change-Id: Ia54101a017b735fb972d58b50c9cae37494872e5