Updated openstack/openstack Project: openstack-infra/shade  a6858ebde096aba991b393db91bf32d9a312bd05
Add test option to use Ansible source repo

This restores the ability to run the Ansible tests using Ansible
from the source repo rather than the production version from pip.
Using production will be the default, but this sets us up to add
a new job to test against the latest dev version, if we want.

Change-Id: I93cdec653dd672acfbc03576d100d19ab8595f2e