Updated openstack/openstack Project: openstack/horizon  227db914bbaeeb9f34f072cd56fe2bd880ac2af6
Change TestImageAdmin to test proper pages

Previously the essence of TestImageAdmin was to do the same things on
same pages, but under a different (admin) user. We realized that it's
not very useful since admin user in Horizon under no circumstances has
less rights than a regular user. A saner approach under admin user
would be to test pages inaccessible to a regular user - that's done in
this patch.

Also is_*_link_available() functions were sligthly improved: they no
longer wait a default timeout for the link to appear, because by the
time these methods are invoked table links are already visible for
sure or are expected to be misssing.

Implements blueprint: horizon-integration-tests-coverage
Change-Id: Idd0c871706d2a8c2c61d6182bbd533bba141baa9