Updated openstack/openstack Project: openstack/oslotest  7f098afd2da15ea007ee707086419d71117028fa
Hack to get back stopall cleanup behavior feature

So, calling addCleanup in __init__ is not supported anymore in
latest testtools. This breaks the support we used to have in
the base class that registered mock.patch.stopall right at the
beginning, so classes inheriting would not have to worry about
the cleanup. For Mitaka, we add a hack to make it work by
looking at the underlying _cleanups variable. This clearly needs
to be removed as soon as possible.

Closes-Bug: #1545576
Change-Id: I29a77224bfd9d106c711155e0cfc10eac0e6fe36