Updated openstack/openstack Project: openstack/ironic  1226bde2ff3ea7e32f1e356e1aa732a2475d2e54
Replace config 'clean_nodes' with 'automated_clean'

Since we have automated cleaning and manual cleaning, the config
'clean_nodes' is a bit confusing since it only refers to automated
cleaning. It is replaced with 'automated_clean'.

This deprecates the 'clean_nodes' config; the plan is to delete it
in Newton cycle.

Updates to the documentation to explicitly mention automated vs
manual cleaning is done in a separate patch.

This includes a reno-style release note about the upgrade impact.

Change-Id: I5b4fbac1753cb4213eb151b663490f381b751597
Partial-Bug: #1526290