Updated openstack/openstack Project: openstack-infra/infra-specs  95613fa92bcbbda3e90c4d7b06bc97354701586b
Update docs-publishing spec to catch up to reality

The openstack-manuals repo no longer has the same constraints as
before and can now fairly easily build all of the manuals on each
change.  Essentially that means that Option 1 has come to pass
(or nearly so).  Update the spec to reflect this.

Change-Id: I596883999e9487676d8ea0fe073bcddd948e4701